chore: moved global middlwares to a dedicated folder
