+++ src/routes/testing.routes.ts
@@ -1,7 +1,7 @@
 import { generateInterviewQuestions, triggerJob } from '#src/controllers/test.controller.js';
 import { sttHandler } from '#src/controllers/test.stt.controller.js';
 import { ttsHandler } from '#src/controllers/test.tts.controller.js';
-import { audioUpload } from '#src/middlewares/multerPosts.js';
+import { audioUpload } from '#src/middlewares/global/multerPosts.js';
 import express, { Router } from 'express';
 
 const router: Router = express.Router();
